requirements.txt
# Core Web Framework
flask==3.0.0
flask-cors==4.0.0

# Data Processing
pandas==2.1.4
pyarrow==14.0.1
numpy==1.26.2

# LLM Integration (choose one or more)
openai==1.6.1
google-generativeai==0.3.2
huggingface-hub==0.20.2

# AWS S3 (for dataset access)
boto3==1.34.10
s3fs==2023.12.2

# Data Generation & Utilities
faker==21.0.0
python-dateutil==2.8.2

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Development Tools
pytest==7.4.3
